requirements.txt
Django==1.9.5
python-dateutil==2.5.2
requests==2.9.1
six==1.10.0
wheel==0.24.0
